refactor(app): split LiveSiteRuntimeFactory out of SiteRuntimeFactory.swift
